LayoutTests/platform/chromium/http/tests/security/contentSecurityPolicy/function-constructor-blocked-expected.txt

CONSOLE MESSAGE: line 12: Uncaught Error: Code generation from strings disallowed for this context
